Added multiple save & protection from deletion of loaned book in inventory

+++ inventory/src/main/java/com/dhitha/lms/inventory/service/InventoryServiceImpl.java
@@ -5,16 +5,19 @@ import com.dhitha.lms.inventory.entity.Inventory;
 import com.dhitha.lms.inventory.error.GenericException;
 import com.dhitha.lms.inventory.error.InventoryNotFoundException;
 import com.dhitha.lms.inventory.repository.InventoryRepository;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Optional;
+import java.util.Set;
+import java.util.UUID;
 import java.util.stream.Collectors;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.util.Assert;
 
 /**
  * Implementation for {@link InventoryService}
@@ -64,40 +67,54 @@ public class InventoryServiceImpl implements InventoryService {
   }
 
   @Override
-  public void add(InventoryDTO inventoryDTO) throws GenericException {
-    try {
-      inventoryRepository.saveAndFlush(mapToEntity(inventoryDTO));
-    } catch (DataIntegrityViolationException e) {
-      log.error(
-          "Error while saving inventory:{} {} ",
-          inventoryDTO.getBookId(),
-          inventoryDTO.getBookReferenceId(),
-          e);
-      throw new GenericException(
-          String.format(
-              "Book with id %s already has reference id %s",
-              inventoryDTO.getBookId(), inventoryDTO.getBookReferenceId()),
-          HttpStatus.BAD_REQUEST.value());
+  @Transactional(rollbackOn = Exception.class)
+  public List<InventoryDTO> add(InventoryDTO inventoryDTO, Integer count) throws GenericException {
+    Assert.notNull(inventoryDTO, "Inventory cannot be null");
+    count = count != null && count > 0 ? count : 1;
+    Set<Inventory> set = new HashSet<>(count);
+    while (count-- > 0) {
+      Inventory inventory = this.mapToEntity(inventoryDTO);
+      inventory.getId().setBookReferenceId(UUID.randomUUID().toString());
+      set.add(inventory);
     }
+    try {
+      inventoryRepository.saveAll(set);
+    } catch (Exception e) {
+      throw new GenericException(
+          "Error while saving inventory", HttpStatus.INTERNAL_SERVER_ERROR.value());
+    }
+    return set.stream().map(this::mapToDTO).collect(Collectors.toList());
   }
 
   @Override
   public void delete(Long bookId) throws InventoryNotFoundException {
-    try {
-      inventoryRepository.deleteByIdBookId(bookId);
-    } catch (EmptyResultDataAccessException e) {
+    List<Inventory> bookInventory = inventoryRepository.findByIdBookId(bookId);
+    if (bookInventory.isEmpty()) {
       throw new InventoryNotFoundException("No inventory found for book id: " + bookId);
     }
+    long count = bookInventory.stream().filter(inventory -> !inventory.getAvailable()).count();
+    if (count > 0) {
+      throw new InventoryNotFoundException(
+          "Some books of book id " + bookId + " are loaned and cannot be deleted ");
+    }
+    inventoryRepository.deleteByIdBookId(bookId);
   }
 
   @Override
   public void delete(Long bookId, String bookReferenceId) throws InventoryNotFoundException {
-    try {
-      inventoryRepository.deleteByIdBookIdAndIdBookReferenceId(bookId, bookReferenceId);
-    } catch (EmptyResultDataAccessException e) {
+    Optional<Inventory> inventory =
+        inventoryRepository.findByIdBookIdAndIdBookReferenceId(bookId, bookReferenceId);
+    if (inventory.isEmpty()) {
       throw new InventoryNotFoundException(
           String.format("No Book found with id %s and reference id %s", bookId, bookReferenceId));
     }
+    if (!inventory.get().getAvailable()) {
+      throw new InventoryNotFoundException(
+          String.format(
+              "Book with id %s and reference id %s is loaned and cannot be deleted",
+              bookId, bookReferenceId));
+    }
+    inventoryRepository.deleteByIdBookIdAndIdBookReferenceId(bookId, bookReferenceId);
   }
 
   private InventoryDTO mapToDTO(Inventory inventory) {
